# api/openapi.yaml
openapi: 3.0.3
info:
  title: Mina REST API
  version: 0.1.0
paths:
...
  /categories:
    get:
      operationId: listCategories
      summary: List categories.
      parameters:
        - name: include_hidden
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: include_tombstoned
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: fqn
            enum:
              - fqn
              - created_at
              - updated_at
        - name: sort_dir
          in: query
          required: false
          schema:
            type: string
            default: asc
            enum:
              - asc
              - desc
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Categories in deterministic FQN order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
    post:
      operationId: createCategory
      summary: Create a category.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /categories/{category_id}:
    parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      operationId: getCategory
      summary: Get a category.
      parameters:
        - name: include_tombstoned
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Category found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateCategory
      summary: Update category hidden state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Category updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCategory
      summary: Tombstone a category.
      responses:
        "204":
          description: Category tombstoned.
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /accounts:
    get:
      operationId: listAccounts
      summary: List accounts.
      parameters:
        - name: include_hidden
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: include_tombstoned
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: fqn
            enum:
              - fqn
              - created_at
              - updated_at
        - name: sort_dir
          in: query
          required: false
          schema:
            type: string
            default: asc
            enum:
              - asc
              - desc
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Accounts in deterministic FQN order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
    post:
      operationId: createAccount
      summary: Create an account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /accounts/{account_id}:
    parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      operationId: getAccount
      summary: Get an account.
      parameters:
        - name: include_tombstoned
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Account found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateAccount
      summary: Update account hidden state and external identifiers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Account updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteAccount
      summary: Tombstone an account.
      responses:
        "204":
          description: Account tombstoned.
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /accounts/{account_id}/credit-limit-history:
    parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      operationId: listCreditLimitHistory
      summary: List credit limit history for an account.
      parameters:
        - name: include_tombstoned
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: effective_date
            enum:
              - effective_date
              - created_at
        - name: sort_dir
          in: query
          required: false
          schema:
            type: string
            default: asc
            enum:
              - asc
              - desc
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Credit limit history in effective-date order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLimitHistoryListResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createCreditLimitHistory
      summary: Create a credit limit history entry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreditLimitHistoryRequest"
      responses:
        "201":
          description: Credit limit history entry created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLimitHistory"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /credit-limit-history/{credit_limit_history_id}:
    parameters:
      - name: credit_limit_history_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      operationId: getCreditLimitHistory
      summary: Get a credit limit history entry.
      parameters:
        - name: include_tombstoned
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Credit limit history entry found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLimitHistory"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCreditLimitHistory
      summary: Tombstone a credit limit history entry.
      responses:
        "204":
          description: Credit limit history entry tombstoned.
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  responses:
    InvalidRequest:
      description: The request is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: The request conflicts with existing state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Account:
      type: object
      additionalProperties: false
      required:
        - account_id
        - fqn
        - kind
        - is_hidden
        - name
        - level
        - created_at
        - updated_at
      properties:
        account_id:
          type: integer
          format: int64
        fqn:
          type: string
        kind:
          type: string
        is_hidden:
          type: boolean
        currency:
          type: string
          nullable: true
          pattern: ^[A-Z]{3}$
        external_id:
          type: string
          nullable: true
        external_system:
          type: string
          nullable: true
        parent_fqn:
          type: string
          nullable: true
        name:
          type: string
        level:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
        tombstoned_at:
          type: string
          nullable: true
    AccountListResponse:
      type: object
      additionalProperties: false
      required:
        - accounts
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/Account"
    CreditLimitHistory:
      type: object
      additionalProperties: false
      required:
        - credit_limit_history_id
        - account_id
        - credit_limit
        - effective_date
        - created_at
      properties:
        credit_limit_history_id:
          type: integer
          format: int64
        account_id:
          type: integer
          format: int64
        credit_limit:
          type: string
          pattern: ^[0-9]{1,18}(\.[0-9]{1,8})?$
          maxLength: 19
        effective_date:
          type: string
          pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
        created_at:
          type: string
        tombstoned_at:
          type: string
          nullable: true
    CreditLimitHistoryListResponse:
      type: object
      additionalProperties: false
      required:
        - credit_limit_history
      properties:
        credit_limit_history:
          type: array
          items:
            $ref: "#/components/schemas/CreditLimitHistory"
...
    APIError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - not_found
            - method_not_allowed
            - conflict
            - internal_error
        message:
          type: string
    Category:
      type: object
      additionalProperties: false
      required:
        - category_id
        - fqn
        - is_hidden
        - name
        - level
        - created_at
        - updated_at
      properties:
        category_id:
          type: integer
          format: int64
        fqn:
          type: string
        is_hidden:
          type: boolean
        parent_fqn:
          type: string
          nullable: true
        name:
          type: string
        level:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
        tombstoned_at:
          type: string
          nullable: true
    CategoryListResponse:
      type: object
      additionalProperties: false
      required:
        - categories
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
    CreateCategoryRequest:
      type: object
      additionalProperties: false
      required:
        - fqn
      properties:
        fqn:
          type: string
        is_hidden:
          type: boolean
          default: false
    CreateAccountRequest:
      type: object
      additionalProperties: false
      required:
        - fqn
      properties:
        fqn:
          type: string
        is_hidden:
          type: boolean
          default: false
        currency:
          type: string
          nullable: true
          pattern: ^[A-Z]{3}$
        external_id:
          type: string
          nullable: true
        external_system:
          type: string
          nullable: true
    CreateCreditLimitHistoryRequest:
      type: object
      additionalProperties: false
      required:
        - credit_limit
        - effective_date
      properties:
        credit_limit:
          type: string
          pattern: ^[0-9]{1,18}(\.[0-9]{1,8})?$
          maxLength: 19
        effective_date:
          type: string
          pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/APIError"
...
    UpdateCategoryRequest:
      type: object
      additionalProperties: false
      required:
        - is_hidden
      properties:
        is_hidden:
          type: boolean
    UpdateAccountRequest:
      type: object
      additionalProperties: false
      required:
        - is_hidden
      properties:
        is_hidden:
          type: boolean
        external_id:
          type: string
          nullable: true
        external_system:
          type: string
          nullable: true
...